Move test classes into sub-namespaces
