+++ tests/Model/LeafTest.php
@@ -1,8 +1,9 @@
 <?php
-namespace Prezly\Slate\Tests;
+namespace Prezly\Slate\Tests\Model;
 
 use Prezly\Slate\Model\Leaf;
 use Prezly\Slate\Model\Mark;
+use Prezly\Slate\Tests\TestCase;
 
 class LeafTest extends TestCase
 {
