+++ tests/Model/MarkTest.php
@@ -1,7 +1,8 @@
 <?php
-namespace Prezly\Slate\Tests;
+namespace Prezly\Slate\Tests\Model;
 
 use Prezly\Slate\Model\Mark;
+use Prezly\Slate\Tests\TestCase;
 
 class MarkTest extends TestCase
 {
