+++ tests/Serialization/fixtures/document_with_data.json
