+++ tests/Serialization/fixtures/document_with_flat_children.json
