+++ tests/Serialization/fixtures/document_with_inline_children.json
