+++ tests/Serialization/fixtures/document_with_nested_children.json
