+++ tests/Serialization/fixtures/document_with_text.json
