+++ tests/Serialization/fixtures/empty_document.json
