+++ tests/Serialization/fixtures/invalid_document_01_not_an_object_int.json
