+++ tests/Serialization/fixtures/invalid_document_02_not_an_object_string.json
