+++ tests/Serialization/fixtures/invalid_document_03_empty_object.json
