+++ tests/Serialization/fixtures/invalid_document_04_invalid_value_object.json
