+++ tests/Serialization/fixtures/invalid_document_07_invalid_document_object.json
