+++ tests/Serialization/fixtures/invalid_document_08_invalid_leaf_object.json
